adding webpack, files .editorconfig, .eslintignore, .eslintrc.yml, .gitignore sepparating project on components
