+++ .editorconfig
@@ -1,0 +1,22 @@
+# Файл с настройками для редактора.
+#
+# Если вы разрабатываете в редакторе от JetBrains, BBEdit, Coda или SourceLair
+# этот файл уже поддерживается и не нужно производить никаких дополнительных
+# действий.
+#
+# Если вы ведёте разработку в другом редакторе, зайдите
+# на http://editorconfig.org и в разделе «Download a Plugin»
+# скачайте дополнение для вашего редактора.
+
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+indent_size = 2
+indent_style = space
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.md]
+trim_trailing_whitespace = false
