+++ .eslintrc.yml
@@ -1,0 +1,10 @@
+env:
+  es6: true
+  browser: true
+  node: true
+
+extends: 'htmlacademy/es6'
+
+parserOptions:
+  sourceType: 'module'
+  ecmaVersion: 2015
