+++ .gitignore
@@ -1,0 +1,16 @@
+# Файлы и папки операционной системы
+.DS_Store
+Thumbs.db
+
+# Файлы редактора
+.idea
+*.sublime*
+.vscode
+
+# Вспомогательные файлы
+*.log*
+node_modules/
+
+# Собранные файлы
+# public/bundle.js
+# public/bundle.js.map
