+++ package.json
@@ -4,12 +4,18 @@
   "private": true,
   "description": "Проект «Киноман» от HTML Academy",
   "scripts": {
+    "build": "webpack",
+    "start": "webpack-dev-server --mode development --open",
     "eslint": "eslint src/**/*.js",
     "test": "npm run eslint"
   },
   "devDependencies": {
+    "clean-webpack-plugin": "4.0.0",
     "eslint": "6.1.0",
-    "eslint-config-htmlacademy": "0.5.2"
+    "eslint-config-htmlacademy": "0.5.2",
+    "webpack": "5.74.0",
+    "webpack-cli": "4.10.0",
+    "webpack-dev-server": "4.9.3"
   },
   "engines": {
     "node": "10"
