+++ public/index.html
@@ -16,7 +16,7 @@
 
   <!-- Звание пользователя -->
 
-  
+
 </header>
 
 <main class="main">
@@ -35,6 +35,6 @@
   </section>
 </footer>
 
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
